backend/requirements.txt
flask>=2.3.0
flask_sqlalchemy>=3.1.0
flask_cors>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
gunicorn>=21.2.0
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=10.0.0
requests>=2.31.0
